Load samples into buffers
This commit adds two new parameter menu entries to load a chosen WAV
file into either of the two Barcode buffers. This makes it easier to
use the script with, for example, something recorded to the internal
norns tape.

When loading an audio file we determine its duration and set the
`buffer_size` accordingly. For consistency with the audio input
recording, we take at most the first 60s of the file.