Allow buffer to be loaded from a file
This commit adds a parameter "audio source". When set to "audio in"
the current behaviour is unchanged (K3 records from the norns audio
input into the currently selected buffer).

When set to "file", the behaviour of K3 is changed. K3 now opens a
file select and the buffer is set to the contents of the file (as
determined by `softcut.buffer_read_mono`). I've used `audio.file_info`
to determine the file length and set `state.buffer_size`
appropriately.

The additional state variable `selecting_file` ensures that the screen
redraw routine called in the lfo metro returns early while the user is
selecting a file.